More work on ORM/schema, pretty-print IPs for Iphdr
Slowly massaging schema into place, playing with things in the console
to make sure things make sense. Also added a pretty-printer for ip_src
and ip_dst in Iphdr as the addresses are just stored as bigints. It sure
will be fun to search based on network addresses! OK, not too bad...