TCP and IP headers mostly done
Dear god, what a pain in the ass! I thought my initial "use 32 dummy
cells at the top and set colspans to bitwidths" solution was a total
hack, but after wasting 2 hours trying to find a more "elegant" way to
do it with CSS, looks like the hack is the right way after all.
Still need to make TCP and IP flags pretty.