Created ip_host_cache table & friends
Created a cache for ip host info as the queries were jsut taking too
long even on my modestly sized DB. Possibly due to my lack of SQL-Fu,
but the cache will definitely speed things up and simplify lots of
stuff. Also created a cache_update table to record time of last update,
and a function refresh_ip_host_cache that refreshes the cache and
updates the timestamp in the cache_update table.